Fix: Move LoginScreen to UI package
